+++ components/hambergurMenu/leftHamMenu.tsx
@@ -250,7 +250,7 @@ const LeftHamMenue = ({
     if (
       responseType === PushResponseType.UploadPostSuccess ||
       responseType === PushResponseType.UploadStorySuccess ||
-      responseType === PushResponseType.AiImageSuccess ||
+      responseType === PushResponseType.AIImageSuccess ||
       responseType === PushResponseType.AIVideoSuccess
     ) {
       return (
@@ -277,7 +277,7 @@ const LeftHamMenue = ({
     if (
       responseType === PushResponseType.UploadPostFailed ||
       responseType === PushResponseType.UploadStoryFailed ||
-      responseType === PushResponseType.AiImageFail ||
+      responseType === PushResponseType.AIImageFailed ||
       responseType === PushResponseType.AIVideoFailed
     ) {
       return (
@@ -306,10 +306,10 @@ const LeftHamMenue = ({
         );
       }
       return "";
-    } else if (notif.ResponseType === PushResponseType.AiImageSuccess && notif.Message) {
+    } else if (notif.ResponseType === PushResponseType.AIImageSuccess && notif.Message) {
       const message = JSON.parse(notif.Message) as IGetMedia;
       return "Your images successfully created by, " + message.version + " model.";
-    } else if (notif.ResponseType === PushResponseType.AiImageFail && notif.Message) {
+    } else if (notif.ResponseType === PushResponseType.AIImageFailed && notif.Message) {
       const message = JSON.parse(notif.Message) as IGetMedia;
       return `Your images failed to be created by " + message.version + " model : ${message.metadata || "Image generation failed."}`;
     } else if (notif.ResponseType === PushResponseType.AIVideoSuccess && notif.Message) {